% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_analysis.R
\name{distance_analysis}
\alias{distance_analysis}
\title{Viewshed Distance Analysis}
\usage{
distance_analysis(
  observer,
  dsm_rast,
  dtm_rast,
  max_distance = 800,
  observer_height = 1.7,
  raster_res = NULL,
  plot = FALSE,
  progress = FALSE,
  cores = 1
)
}
\arguments{
\item{observer}{object of class \code{sf} with POINT geometries; Observer location(s) from where the distance analysis should be computed. It's best to provide a sample of points within the complete study area, to evaluate the Visibility threshold of the study area.}

\item{dsm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DSM}

\item{dtm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DTM}

\item{max_distance}{numeric; Maximum buffer distance}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.7 meters)}

\item{raster_res}{optional; NULL or numeric > 0; Resolution that the viewshed raster should be aggregated to. Must be a multible of the dsm_rast resolution}

\item{plot}{optional; Plot (Visibility ~ Distance)}

\item{progress}{logical; Show progress bar and computation time?}

\item{cores}{numeric; The number of cores to use}
}
\value{
object of class \code{\link[tibble]{tibble}}
}
\description{
Since visibility changes with increasing distance, a threshold for the study area can be evaluated beyond which an observer can't see any terrain.
The \code{distance_analysis} function calculates the proportion of visible area for each distance value.
}
