\name{GVI-package}
\alias{GVI-package}
\alias{GVI}
\docType{package}
\title{
  Green Visibility Index
}
\description{
  This package uses a R and C++ viewshed implementation to calculate a Green Visibility Index (GVI) surface from a DSM, DTM and Greenness Surface.
}
\details{
  As demonstrated by Labib, Huck & Lindley (2021), the visibility of natural greenness is associated with several health benefits along multiple pathways. However, existing methodologies are inadequate for capturing eye-level greenness visibility exposure at high spatial resolutions for observers located on the ground. As a response, we developed a light weighted and fast R package to model and map eye-level greenness visibility exposure at high spatial resolution.
}
\author{
Sebastian Brinkmann, Dr. S.M. Labib

Maintainer: Sebastian Brinkmann <sebastian.brinkmann@fau.de>
}
\references{
   Labib, S., Huck, J. J., & Lindley, S. (2021). Modelling and mapping eye-level greenness visibility exposure using multi-source data at high spatial resolutions. The Science of the total environment, 755 (Pt 1), 143050. https://doi.org/10.1016/j.scitotenv.2020.143050
}
\keyword{ package }
\seealso{
  Python code: https://github.com/jonnyhuck/green-visibility-index
}
