% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoS.R
\name{LoS}
\alias{LoS}
\title{Line of Sight}
\usage{
LoS(rc1, r0, c0, dsm_mat, observerHeight)
}
\arguments{
\item{rc1}{integer vector; Row and column of the targets location in the DSM matrix.}

\item{r0}{integer; Specifying the row of the starting location in the DSM matrix.}

\item{c0}{integer; Specifying the column of the starting location in the DSM matrix.}

\item{dsm_mat}{integer matrix; Matrix derived from a DSM.}

\item{observerHeight}{numeric > 0; Height of the observer (e.g. 1.8)}
}
\value{
Cell numbers of all visible cells of the DSM.
}
\description{
Computes a single Line of Sight (LoS) using a Digital Surface Model (DSM) matrix.
}
