% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LoSCpp}
\alias{LoSCpp}
\title{Line of Sight}
\usage{
LoSCpp(rc1, r0, c0, observerHeight, dsm_mat)
}
\arguments{
\item{rc1}{integer matrix; Row and column of the targets location in the DSM matrix.}

\item{r0}{integer; Specifying the row of the starting location in the DSM matrix.}

\item{c0}{integer; Specifying the column of the starting location in the DSM matrix.}

\item{observerHeight}{numeric > 0; Height of the observer (e.g. 1.7).}

\item{dsm_mat}{numeric matrix; Matrix derived from a DSM.}
}
\value{
Cell numbers of all visible cells of the DSM along the line.
}
\description{
Computes a single Line of Sight (LoS) using a Digital Surface Model (DSM) matrix. Uses a C++ implementation of Bresenham's line algorithm.
}
