% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgvi.R
\name{vgvi}
\alias{vgvi}
\title{Viewshed Greenness Visibility Index (VGVI)}
\usage{
vgvi(
  viewshed,
  sf_start,
  greenspace,
  m = 0.5,
  b = 8,
  mode = c("logit", "exponential")
)
}
\arguments{
\item{viewshed}{object of class \code{\link[terra]{SpatRaster}}; \code{\link[terra]{SpatRaster}} of the \code{\link[GVI]{viewshed}}.}

\item{sf_start}{object of class \code{sf} with one point; Starting location.}

\item{greenspace}{object of class \code{\link[terra]{SpatRaster}}; Binary \code{\link[terra]{SpatRaster}} of the Greenspace mask. Values must be 1 for Green and 0 for No-Green.}

\item{m}{numeric; See ‘Details’.}

\item{b}{numeric; See ‘Details’.}

\item{mode}{character; 'logit' or 'exponential'. See ‘Details’.}
}
\value{
numeric; Weighted VGVI, where 0 = no green cells are visible, and 1 = all of the visible cells are green.
}
\description{
The VGVI expresses the proportion of visible greenness to the total visible area based on a \code{\link[GVI]{viewshed}}.
The estimated VGVI values range between 0 and 1, where 0 = no green cells are visible, and 1 = all of the visible cells are green.
A distance decay function is applied, to account for the reducing visual prominence of an object in space with increasing distance from the observer.
}
\details{
The type of function, used for calculating the distance decay weights, can be defined with the \code{mode} parameter.
The argument 'logit' uses the logistic function, d = 1 / (1 + e^(b * (x - m))) and 'exponential' the exponential function d = 1 / (1 + (b * x^m)).
The decay function can be visualized using the \code{\link[GVI]{visualizeWeights}} function.
}
