% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed.R
\name{viewshed}
\alias{viewshed}
\title{Viewshed}
\usage{
viewshed(
  sf_start,
  dsm_data,
  dtm_data,
  max_distance = 800,
  observer_height = 1.7,
  resolution = NULL,
  plot = FALSE
)
}
\arguments{
\item{sf_start}{object of class \code{sf} with one point; Starting location}

\item{dsm_data}{object of class \code{\link[terra]{SpatRaster}}; \code{\link[terra]{SpatRaster}} of the DSM}

\item{dtm_data}{object of class \code{\link[terra]{SpatRaster}}; \code{\link[terra]{SpatRaster}} of the DTM}

\item{max_distance}{numeric; Buffer distance to calculate the viewshed}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.8)}

\item{resolution}{optional; numeric > 1; Resolution that the GVI should be aggregated to.}

\item{plot}{optional; Plot DSM and GVI}
}
\value{
object of class \code{\link[terra]{SpatRaster}}
}
\description{
Computes the viewshed of a point on a Digital Surface Model map.
}
